==> scripts/04_area_maps.Rmd <==
---
title: "04: Applying the dictionary"
author: "Jamie Montgomery"
output: 
  html_document:
    theme: paper
    toc: true
    toc_float: true
    toc_depth: 2
---
# Summary 

This script takes the geotagged twitter data and applies the dictionary to the tweets to identify whether or not each tweet is considered "nature-based".

## Setup

```{r setup, include=FALSE, message = FALSE, warning = FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(tidyverse)
library(leaflet)
library(sf)
library(mapview)
library(RColorBrewer)
library(ggmap)
library(sp)
library(ggpol) #for the facet_share function
library(grid) #for plot.margin
```

Load data
```{r}
data       <- read_csv("../data/tweets_nature_categorized.csv")
```


```{r}
nature_df <- data %>%
    mutate(coords = gsub("\\)|c\\(", "", geo_coordinates)) %>%
    separate(coords, c("lat", "lon"), sep = ", ") %>%
    mutate_at(c("lon", "lat"), as.numeric) %>% 
    st_as_sf(coords = c("lon", "lat")) %>%
    st_set_crs("+init=epsg:4326") %>%
  mutate(tweet_type = ifelse(nature_word == 1, "nature tweet", "non-nature tweet"),
         nature_user = case_when(
            user_type == "local" & nature_word == 0 ~ "local, non nature tweet",
            user_type == "tourist" & nature_word == 0 ~ "tourist, non nature tweet",
            user_type == "tourist" & nature_word == 1 ~ "tourist, nature tweet",
            user_type == "local" & nature_word == 1 ~ "local, nature tweet"
        ))
```

# Overlap with protected areas

Load california protected areas database
```{r}
cpad <- read_sf("../data/cpad_fixed.shp")
```

Map of CPAD areas in Santa Barbara
```{r}
cpad_map <- mapview(cpad, zcol = "SITE_NAME", legend = FALSE)
cpad_map@map %>% setView(lng = -119.714, lat = 34.426, zoom = 13)
```

**How many tweets come from these areas?**

Count how many points in each polygon (all types of tweets not just nature based)

```{r}
cpad_nature_count <- cpad %>%
  mutate(pt_count = lengths(st_intersects(cpad, nature_df))) %>%
  filter(pt_count > 0)

mapview(cpad_nature_count, zcol = "pt_count")@map %>% setView(lng = -119.714, lat = 34.426, zoom = 13)
```

Save a datafile that identifies the CPAD area for each tweet

```{r}
tweets_cpad <- st_intersection(nature_df, cpad)

write_sf(tweets_cpad, "../data/tweets_in_cpad_areas.shp")
```


Some of these areas overlap with high tourist areas (e.g. the Zoo and Wharf). Let's look at the content of these tweets to see how often tweets coming from these locations are nature-based or not.

# Compare occurrence of nature vs non-nature based tweets

```{r}
nature_tweets <- nature_df %>%
  filter(nature_word == 1)
non_nature_tweets <- nature_df %>% 
  filter(nature_word == 0)

cpad_all_count <- cpad %>%
  mutate(total_tweets = lengths(st_intersects(cpad, nature_df)),
         nature_count = lengths(st_intersects(cpad, nature_tweets)),
         non_nature_count = lengths(st_intersects(cpad, non_nature_tweets))) %>%
  rowwise() %>%
  mutate(ratio = nature_count/non_nature_count,
         prop  = nature_count/sum(c(nature_count, non_nature_count))) %>%
  filter(!is.na(ratio)) %>%
  mutate(ratio = ifelse(is.infinite(ratio), nature_count, ratio)) %>% 
  st_set_geometry("geometry")
```


The highest ratio of nature tweets to non-nature takes place at `r filter(cpad_all_count, ratio == max(cpad_all_count$ratio))$SITE_NAME`.

```{r}
m <- mapview(cpad_all_count %>% select(prop, ratio, total_tweets), burst = TRUE, hide = TRUE) +
  mapview(nature_df %>% filter(nature_word == 1))


m@map %>% setView(lng = -119.714, lat = 34.426, zoom = 13)
```


Let's look at the top 20 most popular sites. Names in **bold** indicate over 50% of tweets are nature-based

```{r}
top_20 <- cpad_all_count %>%
  st_set_geometry(NULL) %>%
  arrange(-nature_count) %>%
  slice(1:20) %>%
  pivot_longer(cols = c(nature_count, non_nature_count), names_to = "tweet_type", values_to = "count")

top_20_prop <- top_20 %>% 
  select(SITE_NAME, prop, count) %>%
  group_by(SITE_NAME, prop) %>%
  summarise(count = sum(count)) %>%
  ungroup() %>%
  distinct() %>%
  mutate(prop = paste0(round(100*prop,0),"% "))

ggplot(top_20, aes(x = reorder(SITE_NAME, count), fill = tweet_type, y = count)) +
  geom_bar(stat = "identity", position = "stack") +
  theme_minimal() +
  coord_flip() +
  labs(x = "",
       y = "Number of tweets",
       fill = "",
       title = "Top 20 most tweeted from protected areas* in Santa Barbara") +
  scale_fill_manual(values = c("darkgreen", "gray"), labels = c("Nature-based", "Other")) +
  theme(legend.position = "none",
        plot.title = element_text(hjust = 3.3),
        axis.text.y = element_text(face = rev(c('plain', 'bold', 'plain', 'plain', 'bold', 'bold', 'bold', 'plain', 'plain', 'bold','bold','bold','bold','bold','bold','bold','bold','plain','bold','bold')))) +
    geom_text(aes(SITE_NAME, y = count, label = prop, fill = NULL), 
              data = top_20_prop, hjust = -0.05, size = 3) +
  ylim(0, 1510)
```

What's going on at Manning Park? A relatively small park in Montecito.

```{r}
manning_park <- filter(nature_df, geo_coordinates %in% c("c(34.43361111, -119.63194444)", "c(34.4337, -119.632)"))

kable(sample_n(manning_park %>% select(date, full_text, user_location, user_type, nature_word), 7)) %>%
  kable_styling(bootstrap_options = c("striped", "condensed"), font_size = 11, fixed_thead = T)

```

Given the low proportion of tweets that are nature based, I think this is a default coordinate for "Montecito, California" when geo-tagging.

Do the same but with parks that have at least two tweets. We get an oddity here with places that have just 1 tweet and that is nature-based (100%)

```{r}
top_20 <- cpad_all_count %>%
  group_by(SITE_NAME) %>%
  mutate(total_tweets = sum(nature_count, non_nature_count)) %>%
  ungroup() %>%
  filter(total_tweets > 1) %>%
  st_set_geometry(NULL) %>%
  arrange(-prop) %>%
  slice(1:20) 

ggplot(top_20, aes(x = reorder(SITE_NAME, prop), y = prop)) +
  geom_bar(stat = "identity", fill = "darkgreen") +
  theme_minimal() +
  coord_flip() +
  labs(x = "",
       y = "Proportion of tweets that are nature-based",
       title = "Top 20 sites with highest proportion of nature-based tweets") +
  theme(plot.title = element_text(hjust = 20)) 
```

## How does this differ across tourists and locals?

```{r}
cpad_locals <- cpad %>%
  mutate(nature_count = lengths(st_intersects(cpad, nature_tweets %>% filter(user_type == "local"))),
         non_nature_count = lengths(st_intersects(cpad, non_nature_tweets %>% filter(user_type == "local")))) %>%
  rowwise() %>%
  mutate(
         ratio = nature_count/non_nature_count,
         prop  = nature_count/sum(c(nature_count, non_nature_count))) %>%
  filter(!is.na(ratio)) %>%
  mutate(ratio = ifelse(is.infinite(ratio), nature_count, ratio),
         user_type = "local") %>% 
  st_set_geometry("geometry")

cpad_tourists <- cpad %>%
  mutate(nature_count = lengths(st_intersects(cpad, nature_tweets %>% filter(user_type == "tourist"))),
         non_nature_count = lengths(st_intersects(cpad, non_nature_tweets %>% filter(user_type == "tourist")))) %>%
  rowwise() %>%
  mutate(
         ratio = nature_count/non_nature_count,
         prop  = nature_count/sum(c(nature_count, non_nature_count))) %>%
  filter(!is.na(ratio)) %>%
  mutate(ratio = ifelse(is.infinite(ratio), nature_count, ratio),
         user_type = "tourist") %>% 
  st_set_geometry("geometry")

cpad_users <- cpad_locals %>%
  rbind(cpad_tourists)
```

Look at proportion and number of tweets by locals
```{r locals_plots}

top_20_local <- cpad_locals %>%
  st_set_geometry(NULL) %>%
  arrange(-nature_count) %>%
  slice(1:20) %>%
  pivot_longer(cols = c(nature_count, non_nature_count), names_to = "tweet_type", values_to = "count")

top_20_prop_local <- top_20_local %>% 
  select(SITE_NAME, prop, count) %>%
  group_by(SITE_NAME, prop) %>%
  summarise(count = sum(count)) %>%
  ungroup() %>%
  distinct() %>%
  mutate(prop = paste0(round(100*prop,0),"% "))

locals_prop <- ggplot(top_20_local, aes(x = reorder(SITE_NAME, count), fill = tweet_type, y = count)) +
  geom_bar(stat = "identity", position = "stack") +
  theme_minimal() +
  coord_flip() +
  labs(x = "",
       y = "Number of tweets",
       fill = "",
       title = "Top 20 most tweeted from protected areas* by locals in Santa Barbara") +
  scale_fill_manual(values = c("darkgreen", "gray"), labels = c("Nature-based", "Other")) +
  theme(legend.position = "none",
        plot.title = element_text(hjust = 1.8)) +
    geom_text(aes(SITE_NAME, y = count, label = prop, fill = NULL), 
              data = top_20_prop_local, hjust = -0.05, size = 3) +
  ylim(0, 800)
locals_prop
```

```{r tourists_plots}

top_20_tourist <- cpad_tourists %>%
  st_set_geometry(NULL) %>%
  arrange(-nature_count) %>%
  slice(1:20) %>%
  pivot_longer(cols = c(nature_count, non_nature_count), names_to = "tweet_type", values_to = "count")

top_20_prop_tourist <- top_20_tourist %>% 
  select(SITE_NAME, prop, count) %>%
  group_by(SITE_NAME, prop) %>%
  summarise(count = sum(count)) %>%
  ungroup() %>%
  distinct() %>%
  mutate(prop = paste0(round(100*prop,0),"% "))

tourists_prop <- ggplot(top_20_tourist, aes(x = reorder(SITE_NAME, count), fill = tweet_type, y = count)) +
  geom_bar(stat = "identity", position = "stack") +
  theme_minimal() +
  coord_flip() +
  labs(x = "",
       y = "Number of tweets",
       fill = "",
       title = "Top 20 most tweeted from protected areas* by tourists in Santa Barbara") +
  scale_fill_manual(values = c("darkgreen", "gray"), labels = c("Nature-based", "Other")) +
  theme(legend.position = "none",
        plot.title = element_text(hjust = 1.8)) +
    geom_text(aes(SITE_NAME, y = count, label = prop, fill = NULL), 
              data = top_20_prop_tourist, hjust = -0.05, size = 3) +
  ylim(0, 800) 

tourists_prop
```

Areas with at least 25 tweets.

```{r, message = FALSE, warning= FALSE}
pyramid_df <- cpad_users %>%
  st_set_geometry(NULL) %>%
  #mutate(prop_non_nature = 1-prop) %>%
  pivot_longer(cols = c(nature_count, non_nature_count), names_to = "tweet_type", values_to = "count") %>%
  group_by(SITE_NAME) %>%
  mutate(total = sum(count)) %>%
  ungroup() %>%
  filter(total > 24) %>%
  mutate(count = ifelse(user_type == "local", -1*count, count),
         prop = ifelse(tweet_type == "non_nature_count", NA, prop))

labs <- c(local = "Locals", tourist = "Tourists")

ggplot(pyramid_df, aes(x = reorder(SITE_NAME, total), fill = tweet_type, y = count)) +
  geom_bar(stat = "identity", position = "stack") +
  #geom_text(aes(label = round(prop*100,0), y = count), size = 3) +
  facet_share(~user_type, dir = "h", scales = "free", reverse_num = TRUE, 
              labeller = labeller(user_type = labs)) +
  coord_flip() +
  theme_minimal() +
  scale_fill_manual(values = c("darkgreen", "gray"), labels = c("Nature-based", "Other")) +
  theme(legend.position = "none",
        plot.margin=grid::unit(c(0,10,0,-60), "mm")) +
  labs(x=NULL,y=NULL)

ggsave("../figs/pyramid_plot_num_tweets_per_area.png")
```

Scatterplot of this same data

```{r}
scatter_df <- pyramid_df %>%
  select(SITE_NAME, prop, user_type, tweet_type, total) %>%
  filter(tweet_type == "nature_count") %>%
  pivot_wider(names_from = user_type, values_from = prop) %>%
  mutate(
    label = ifelse(
      SITE_NAME %in% c("Arroyo Burro Beach County Park", "Leadbetter Beach", "Douglas Family Reserve", "Rattlesnake Canyon Park", "Los Padres National Forest", "Mission Park", "Alameda Park","Cachuma Lake Recreation Area", "Laurel Canyon Reservoir", "Twin Lakes Golf Course"), as.character(SITE_NAME), ""),
    fill = ifelse(label == "", 0, 1))

ggplot(scatter_df, aes(x = local, y = tourist)) +
  geom_point(aes(color = fill, size = total)) +
  geom_abline(color = "gray") +
  theme_classic() +
  theme(legend.position = "none") +
  geom_text_repel(aes(label = label), size = 2.85) +
  labs(x = "Locals", y = "Tourists", title = "Proportion of nature-based tweets by locals and tourists \nat Santa Barbara protected areas")

```

## Do tourists and locals visit the same or different nature sites?

At the lower end we see more locals than tourists visiting these sites. These tend to be less popular areas. On the upper end, we see sites that are more frequented overall, and more frequented by tourists. These include well-known areas like the Santa Barbara Harbor and Stearn's Wharf. Those on the lower end that locals frequent more are either lesser-known (Shoreline Park, Alameda Park are both neighborhood parks), or further from main tourist areas (e.g. Goleta Beach)

```{r}
tweets_in_cpad <- read_sf("../data/tweets_in_cpad_areas.shp")

twt_df <- tweets_in_cpad %>%
  select(user_id, usr_typ, SITE_NA) %>%
  st_set_geometry(NULL) %>%
  distinct() %>%
  group_by(SITE_NA, usr_typ) %>%
  summarize(count = n()) %>%
  ungroup() %>%
  group_by(usr_typ) %>%
  mutate(total_users = sum(count)) %>%
  ungroup() %>%
  mutate(prop = count/total_users) %>%
  select(-total_users, -count) %>%
  pivot_wider(names_from = usr_typ, values_from = prop) %>%
  mutate(local = ifelse(is.na(local), 0, local),
         tourist = ifelse(is.na(tourist), 0, tourist),
    label = ifelse(
      SITE_NA %in% c("Arroyo Burro Beach County Park", "Santa Barbara Harbor", "Stearns Wharf", "Los Padres National Forest", "Carpinteria State Beach", "Santa Barbara Bowl", "Santa Barbara Maritime Museum", "Manning Park", "Santa Barbara Zoological Gardens", "Chase Palm Park"), as.character(SITE_NA), ""),
    fill = ifelse(label == "", "fill", "empty"))
  
ggplot(twt_df, aes(x = local, y = tourist)) +
  geom_point(aes(color = fill), size = 2) +
  scale_color_manual(values = c("black", "gray80")) +
  geom_abline(color = "gray") +
  theme_classic() +
  theme(legend.position = "none") +
  geom_text_repel(aes(label = label), size = 2.75) +
  labs(x = "Locals", y = "Tourists", title = "Proportion of users visiting Santa Barbara protected areas")

```

Zoom in on lower end

```{r}

lower_end <- twt_df %>% 
  filter(local < 0.04 & tourist < 0.04) %>%
  mutate(local = ifelse(is.na(local), 0, local),
         tourist = ifelse(is.na(tourist), 0, tourist),
    label = ifelse(
      SITE_NA %in% c("Cachuma Lake Recreation Area", "Leadbetter Beach", "Goleta Beach County Park", "Shoreline Park", "Coal Oil Point Reserve", "Alameda Park", "Douglas Family Reserve", "El Presidio de Santa Barbara State Historic Park"), as.character(SITE_NA), ""),
    fill = ifelse(label == "", "fill", "empty"))

ggplot(lower_end, aes(x = local, y = tourist)) +
  geom_point(aes(color = fill), size = 2) +
  scale_color_manual(values = c("black", "gray80")) + 
  geom_abline(color = "gray") +
  theme_classic() +
  theme(legend.position = "none") +
  geom_text_repel(aes(label = label), size = 2.75) +
  labs(x = "Locals", y = "Tourists", title = "Proportion of users visiting Santa Barbara protected areas")
```


We only have 5,484 unique users from with CPAD areas. This is X% of all unique users in SB.

## What sites have no nature-based tweets?

```{r}
cpad_users %>% 
  filter(nature_count == 0) %>%
  select(SITE_NAME) %>%
  mapview(zcol = "SITE_NAME", legend = FALSE)
```

## What areas of Santa Barbara have over 50% nature-based tweets but are not within a designated CPAD area?

We need to remove the tweets that are within CPAD areas first.

```{r}
hex_grid <- read_sf("../data/sb_area_hexagons_larger.shp")

#get all tweets not in a CPAD area
tweets_nin_cpad <- nature_df %>%
  mutate(cpad = lengths(st_intersects(nature_df, cpad))) %>%
  filter(cpad == 0) %>%
  select(-cpad)

nature_tweets <- tweets_nin_cpad %>%
  filter(nature_word == 1)
non_nature_tweets <- tweets_nin_cpad %>% 
  filter(nature_word == 0)

#overlap the hex grids with tweets
hex_tweet_counts <- hex_grid %>%
  mutate(total_tweets = lengths(st_intersects(hex_grid, tweets_nin_cpad)),
         nature_count = lengths(st_intersects(hex_grid, nature_tweets)),
         non_nature_count = lengths(st_intersects(hex_grid, non_nature_tweets))) %>%
  rowwise() %>%
  mutate(ratio = nature_count/non_nature_count,
         prop  = nature_count/sum(c(nature_count, non_nature_count))) %>%
  filter(!is.na(ratio)) %>%
  mutate(ratio = ifelse(is.infinite(ratio), nature_count, ratio)) %>% 
  st_set_geometry("geometry")
```


```{r}
m <- mapview(hex_tweet_counts, zcol = "prop") +
  mapview(cpad, zcol = "SITE_NAME", legend = FALSE) 

m@map %>% setView(lng = -119.714, lat = 34.426, zoom = 13)
```


### Percentage of tweets that are nature-based across Santa Barbara

```{r, fig.width = 9}

hex_grid <- read_sf("../data/sb_area_hexagons.shp")


#overlap the hex grids with tweets
hex_tweet_counts_nature <- hex_grid %>%
  mutate(nature_twts = lengths(st_intersects(hex_grid, nature_df %>% filter(nature_word == 1))),
         non_nature_twts = lengths(st_intersects(hex_grid, nature_df %>% filter(nature_word == 0))),
         total = lengths(st_intersects(hex_grid, nature_df))) %>%
  rowwise() %>%
  mutate(prop = round(nature_twts/total*100),2) %>%
  filter(total > 0) %>% 
  st_set_geometry("geometry")


m <- mapview(hex_tweet_counts_nature, zcol = "prop", 
             col.regions = colorRampPalette(brewer.pal(9, "Greens")[5:9]), alpha = 0.9,
             layer.name = "Percentage (%)")

m@map %>% setView(lng = -119.714, lat = 34.426, zoom = 13)
```

------

Does this hold true over time?

Create a plot that shows proportion of nature-based tweets by site over time (might be too many, I would remove los padres forest)

```{r}

out <- data.frame()
for(i in 2015:2019){
  print(i)
  yr_tweets <- nature_tweets %>% filter(Year == i)
  yr_tweets_nn <- non_nature_tweets %>% filter(Year == i) #non-nature
  
cpad_by_year <- cpad %>%
  st_set_geometry(NULL) %>%
  mutate(nature_count = lengths(st_intersects(cpad, yr_tweets)),
         non_nature_count = lengths(st_intersects(cpad, yr_tweets_nn))) %>%
  rowwise() %>%
  mutate(
         ratio = nature_count/non_nature_count,
         prop  = nature_count/sum(c(nature_count, non_nature_count)),
         year = i) %>%
  filter(!is.na(ratio)) %>%
  mutate(ratio = ifelse(is.infinite(ratio), nature_count, ratio))

out <- rbind(out, cpad_by_year)

}

yrly_df <- out %>%
  group_by(SITE_NAME) %>%
  mutate(num_years = n()) %>%
  ungroup() %>%
  mutate(total_tweets = nature_count + non_nature_count)

#remove sites with just one year of tweets

ggplot(yrly_df %>% filter(num_years>1), aes(x = year, y = prop, color = SITE_NAME)) +
  geom_path(aes(size = nature_count), lineend="round") +
  theme_minimal() +
  theme(legend.position = "none") +
  facet_wrap(~SITE_NAME)

ggplot(yrly_df %>% filter(num_years>1), aes(x = year, y = prop)) +
  geom_col(aes(fill = total_tweets)) +
  theme_minimal() +
  theme(legend.position = "bottom") +
  facet_wrap(~SITE_NAME) +
  labs(x = "",
       y = "Proportion tweets nature-based")
```
